console: Add 'k' option to exit on console-down
This commit is adding an option 'k' to make the console
application exits when the console is down.

This can be useful in some cases such as an integration to LAVA
infrastructure. Console can be used to open a terminal on different
boards to be able to run some tests.
In some cases, the console is down and without this commit, the
tests will be timed-out because it can't talk to the device.
Adding this option will allow us to exit the console directly, without
waiting for a timeout. The benefit will be time saving.

Signed-off-by: Mylène Josserand <mylene.josserand@collabora.com>